ROOT revision 41453 caused TMath::Power() to be ambiguously overloaded when the second arg was an unsigned int.  Problem reported and rejected by ROOT:   https://savannah.cern.ch/bugs/index.php?87978 Cast 2nd arg to "int" is both forward/backward ROOT compatible.
git-svn-id: file:///home/mroda/Software/Genie/NewGlobal/trunk@3361 cc9776de-3512-45ca-aafc-e2d9ed43c22c

+++ src/Numerical/Simpson2D.cxx
@@ -107,13 +107,13 @@ double Simpson2D::Integrate(GSFunc & gsfunc) const
     int idim=-1;
     if(fFastDensityIncrease) {
       // increase the grid density fast - all dimensions simultaneously
-      np = (unsigned int) TMath::Power(2,n) + 1;
+      np = (unsigned int) TMath::Power(2,(int)n) + 1;
       n++;
       fmap.IncreaseGridDensity(np);
     } else {
       // increase the grid density slowly - 1 dimension at a time...
       if(iter%ndim==0) {
-        np = (unsigned int) TMath::Power(2,n) + 1;
+        np = (unsigned int) TMath::Power(2,(int)n) + 1;
         n++;
         idim = 0;
       }
